So much has happened since my last push. This will be a long comment.
I am pleased with the fact that I've been able to refactor the code so that I can complete a training run locally and log data with wandb

I am very pleased that the data gathered by wandb is extensive, and i believe that this will ultimately unlock numerous vectors for model optimization.

There are numerous problems with this code. I can write a list, however the list is too long to post here.

In addition, there are shortcomings with the code. The shortcomings could largely be dealt with by selectively including functions from 2x.

In spite of these deficiencies, the code performs surprisingly well. I've run inference and achieved 0.630 compared to my own 0.670 using Abishek's approach, and 0.690 using abishek's checkpoints.

There is value here.

The huggingface approach is unsurprisingly heavily dependent on abstracting away "base PyTorch" and using transformers methods and associated libraries. This is not surprising, however it is limiting in some cases. One obvious case is Kfolds. There is no clear method for using Kfolds in the context of a model built around these libraries. I believe that the inclusion of Kfolds will improve performance. There are other examples.

I currently have the following views about the final 14 days:
- the core of the cod e should be base pytorch. minimal abstraction
- the methods and functions from 2x Longformers (Abishek's code) can be transplanted into the new model.
- The methods and functions from the huggingface approach which are useful can be transposed into the new model (after fixing the problems). This will facilitate the NLP data tasks like spelling etc, post-processing, tokenization tuning etc.
- integration with wandb will allow hyperparameter optimization

I don't know if we can do this over the next 14 days, but I'm damn well going to try. The main help i need from you is to do the refactor we talked about to get longformer x2 working without tez, in base pytorch. I think I can (and should) handle the rest.