have a look at hf_functions.py, line 154 - 173, fix_beginnings() function. This is an example of what i mentioned last night. labels is an input parameter, and its also the return/output.
prima facie it seems self-referential adn circular. Maybe that's not a problem in terms of programming, but at the very least i find it a little confusing. Would it not be better to have some differentiation? eg: the input param called 'old_labels' and the return/output called 'new_labels'?

there are other examples but this one is really obvious.