I think the problem is still in cell 8. The error doesn't appear here, but I believe it originates here.
At first, based on the error message, I thought that the issue was that predictions were not getting appended to the list. A simple check disproved that.

Then i noticed that there are only 9 TQDM progress bars. There should be more. I went back and checked the ensemble of 2 models that succeeds, and there are 10 TQDM progress bars. This makes sense as there are a total of 10 folds or checkpoints.

In the new notebook, I would therefore expect a total of 20 tqdm progress bars from cell 8 sa the ensemble of 4 models as 20 folds or checkpoints. The fact that there are only 9 leads me ot beleive that this is where the problem is.