I'll be damned. There is something to be said for trial and error. Or maybe its just sheer bloody minded persistence.
__feedback_prize_2021_ensembling.ipynb now executes without error on all 20 checkpoints from all 4 trained models.

Can't wait to give this a run on Kaggle. Interesting but it runs so much more slowly on Kaggle. Its taken 10 minutes to execute the last version which used 15 checkpoints from 3 models, and much longer to score. I want to see the score on that, then I'll update to use this code and score on 20 checkpoints.

Regardless of scores, all this is still "naive" so I expect that once we can add the bayesian optimisation to the ensemble the score will improve.

Whether this all comes together in time for end of competition is another thing entirely...